Move group implementations to own module
